+++ src/Regressors/MLPRegressor.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\Regressors;
 
 use Rubix\ML\Online;
+use Rubix\ML\Verbose;
 use Rubix\Tensor\Matrix;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Dataset;
@@ -11,6 +12,7 @@ use Rubix\ML\Datasets\DataFrame;
 use Rubix\ML\NeuralNet\Snapshot;
 use Rubix\ML\NeuralNet\FeedForward;
 use Rubix\ML\NeuralNet\Layers\Hidden;
+use Rubix\ML\Other\Traits\LoggerAware;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
 use Rubix\ML\NeuralNet\Layers\Continuous;
 use Rubix\ML\NeuralNet\Layers\Placeholder;
@@ -35,8 +37,10 @@ use RuntimeException;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class MLPRegressor implements Online, Persistable
+class MLPRegressor implements Online, Verbose, Persistable
 {
+    use LoggerAware;
+
     /**
      * The user-specified hidden layers of the network.
      *
@@ -301,6 +305,8 @@ class MLPRegressor implements Online, Persistable
             return;
         }
 
+        !isset($this->logger) ?: $this->logger->info('Training started');
+
         $n = $dataset->numRows();
 
         list($testing, $training) = $dataset->stratifiedSplit($this->holdout);
@@ -311,7 +317,7 @@ class MLPRegressor implements Online, Persistable
         $bestSnapshot = null;
         $previous = INF;
 
-        for ($epoch = 0; $epoch < $this->epochs; $epoch++) {
+        for ($epoch = 1; $epoch <= $this->epochs; $epoch++) {
             $batches = $training->randomize()->batch($this->batchSize);
 
             $loss = 0.;
@@ -331,6 +337,9 @@ class MLPRegressor implements Online, Persistable
                 $bestScore = $score;
                 $bestSnapshot = Snapshot::take($this->network);
             }
+
+            !isset($this->logger) ?: $this->logger->info("Epoch $epoch"
+                . " completed, loss: $loss");
 
             if ($score === $max) {
                 break 1;
@@ -355,8 +364,13 @@ class MLPRegressor implements Online, Persistable
         if (end($this->scores) < $bestScore) {
             if (isset($bestSnapshot)) {
                 $this->network->restore($bestSnapshot);
+
+                !isset($this->logger) ?: $this->logger->info('Network restored'
+                    . ' from previous snapshot');
             }
         }
+
+        !isset($this->logger) ?: $this->logger->info("Training complete");
     }
 
     /**