+++ tests/Clusterers/FuzzyCMeansTest.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\Tests\Clusterers;
 
 use Rubix\ML\Learner;
+use Rubix\ML\Verbose;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Probabilistic;
@@ -39,6 +40,7 @@ class FuzzyCMeansTest extends TestCase
         $this->assertInstanceOf(FuzzyCMeans::class, $this->estimator);
         $this->assertInstanceOf(Learner::class, $this->estimator);
         $this->assertInstanceOf(Probabilistic::class, $this->estimator);
+        $this->assertInstanceOf(Verbose::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
     }