+++ tests/Regressors/MLPRegressorTest.php
@@ -4,6 +4,7 @@ namespace Rubix\ML\Tests\Regressors;
 
 use Rubix\ML\Online;
 use Rubix\ML\Learner;
+use Rubix\ML\Verbose;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Unlabeled;
@@ -47,6 +48,7 @@ class MLPRegressorTest extends TestCase
         $this->assertInstanceOf(MLPRegressor::class, $this->estimator);
         $this->assertInstanceOf(Online::class, $this->estimator);
         $this->assertInstanceOf(Learner::class, $this->estimator);
+        $this->assertInstanceOf(Verbose::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
     }