+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Unreleased
     - Added Verbose interface for logging
+    - Estimators with explicit loss functions are now Verbose
     - Grid Search: Added option to retrain best model on full dataset
 
 - 0.0.6-alpha